move createInitialState from frontend folder, add eslintrc for server
