+++ gulp/native.js
@@ -3,7 +3,7 @@ import gulp from 'gulp';
 
 // Native can't run Node code, so we need gulp task to pre-write initialState.
 gulp.task('native', ['env'], (done) => {
-  const createInitialState = require('../src/server/frontend/createInitialState').default;
+  const createInitialState = require('../src/server/createInitialState').default;
 
   const initialState = createInitialState();
   const string = JSON.stringify(initialState, null, 2);
